+++ client/react-utils/components/auto-save-field/action-creators.js
@@ -1,0 +1,4 @@
+import actionCreator from './../../action-creator';
+import actions from './actions';
+
+export const updateStatus = (key, status) => actionCreator(actions.UPDATE_STATUS, { key, status });