+++ client/react-utils/components/auto-save-field/namespace.js
@@ -1,0 +1,3 @@
+import namespace from './../namespace';
+
+export default (key) => namespace(`AUTO_SAVE_FIELD.${key}`);