+++ client/react-utils/components/filter-input/actions.js
@@ -1,0 +1,6 @@
+import namespace from './namespace';
+import namespaceKeys from './../../namespace-keys';
+
+export default namespaceKeys(namespace, [
+    'UPDATE_VALUE'
+]);