+++ client/react-utils/components/namespace.js
@@ -1,0 +1,4 @@
+import namespace from './../namespace';
+import packageJson from './../../../package.json';
+
+export default (key) => namespace(`${packageJson.name}.${key}`);