+++ client/react-utils/components/reducers.js
@@ -1,0 +1,8 @@
+import autoSaveFieldReducers from './auto-save-field/reducers';
+import filterInputReducers from './filter-input/reducers';
+import concatenateReducers from './../concatenate-reducers';
+
+export default concatenateReducers([
+    autoSaveFieldReducers,
+    filterInputReducers
+]);