+++ client/react-utils/components/selector.js
@@ -1,0 +1,4 @@
+import packageJson from './../../../package.json';
+import selector from './../selector';
+
+export default (state, Component, id) => selector(state, packageJson.name, Component, id);