+++ client/react-utils/namespace-base.js
@@ -1,4 +1,0 @@
-import namespace from './namespace';
-import packageJson from './../../package.json';
-
-export default (key) => namespace(`${packageJson.name}.${key}`);