+++ client/react-utils/selector-base.js
@@ -1,4 +1,0 @@
-import packageJson from './../../package.json';
-import selector from './selector';
-
-export default (state, Component, id) => selector(state, packageJson.name, Component.displayName, id);